backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
itsdangerous>=2.1.2
Pillow>=9.0.0
numpy>=1.20.0
rasterio>=1.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
motor>=3.3.0
boto3>=1.34.0
tqdm>=4.65.0
